backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
langchain
langchain-openai
langchain-experimental
langchain-community
openai
dateparser
